CMakeDeps backporting future change from CMake newer versions. Removed PkgConfigDeps change (too risky). Moved and modified test
